refactor(sidepanel): extract storage module (Phase 2/3)
